+++ csrc/quickjs.h
@@ -1052,7 +1052,7 @@ int JS_SetModuleExport(JSContext *ctx, JSModuleDef *m, const char *export_name,
 int JS_SetModuleExportList(JSContext *ctx, JSModuleDef *m,
                            const JSCFunctionListEntry *tab, int len);
 JS_BOOL JS_IsUint8Array(JSValueConst v);
-uint8_t* JS_Uint8ArrayGetBuffer(JSValueConst v, uint32_t *size);
+uint8_t* JS_Uint8ArrayGetBuffer(JSValueConst v, size_t *size);
 JSValue JS_NewUint8Array(JSContext *ctx, const uint8_t* data, uint32_t data_len);
 
 enum {
